feat(settings): Drop public/private settings
Use django-environ to split settings up into categories
and install defaults.  Use .env file in root to handle
settings we dont want and use an .env.dev file locally

Aside from aws credentials you can test without
any private settings
