+++ .env.example
@@ -1,0 +1,68 @@
+DEBUG=on
+DEVELOPMENT=on
+TESTING=on
+
+AWS_ACCESS_KEY_ID=***
+AWS_SECRET_ACCESS_KEY=***
+
+SENTRY_DSN=***
+
+SOLR_HOST=***
+SOLR_RECAP_HOST=***
+SOLR_PAGERANK_DEST_DIR=***
+
+REDIS_HOST=***
+REDIS_PORT=***
+
+DOCKER_SELENIUM_HOST=***
+DOCKER_DJANGO_HOST=***
+
+SELENIUM_HEADLESS=False
+
+DOCTOR_HOST=***
+DISCLOSURE_HOST=***
+SECRET_KEY=********
+
+MAILCHIMP_SECRET=***
+MAILCHIMP_API_KEY=***
+
+# PACER
+PACER_USERNAME=***
+PACER_PASSWORD=***
+
+
+EIN_SECRET=***
+TIME_ZONE="America/***"
+FUNDRAISING_MODE=False
+
+# Internet Archive
+IA_ACCESS_KEY=***
+IA_SECRET_KEY=***
+IA_COLLECTIONS=[]
+IA_OA_COLLECTIONS=***
+
+
+#Postgres
+PG_USER=***
+PG_PASSWORD=***
+
+#ALLOWED HOSTS=***,***,***
+
+#GENERAL KEYS
+PAYPAL_ENDPOINT=https://api.paypal.com
+PAYPAL_ACCOUNT=*******
+PAYPAL_CLIENT_ID=""
+PAYPAL_SECRET_KEY=""
+STRIPE_SECRET_KEY=""
+STRIPE_PUBLIC_KEY=""
+
+
+CL_API_TOKEN=*****
+
+LASC_USERNAME=****
+LASC_PASSWORD=****
+
+ADMIN_NAME=****
+ADMIN_EMAIL=****
+
+API_READ_DATABASES=****
