+++ .github/workflows/tests.yml
@@ -41,9 +41,6 @@ jobs:
         uses: actions/checkout@v2-beta
         with:
           path: courtlistener
-      - name: Create the private settings file
-        working-directory: courtlistener/cl/settings
-        run: cp 05-private.example 05-private.py
 
       - name: Echo github actor name for debugging
         run: echo ${{ github.actor }}
