+++ docker/courtlistener/docker-compose.yml
@@ -58,9 +58,8 @@ services:
       - ${CL_BASE_DIR:-../../../courtlistener}:/opt/courtlistener
     networks:
       - cl_net_overlay
-    environment:
-      AWS_ACCESS_KEY_ID: ${AWS_ACCESS_KEY_ID}
-      AWS_SECRET_ACCESS_KEY: ${AWS_SECRET_ACCESS_KEY}
+    env_file:
+      - ../../.env.dev
 
   cl-django:
     container_name: cl-django
@@ -81,9 +80,8 @@ services:
       - ${CL_BASE_DIR:-../..}/cl/assets/media:/opt/courtlistener/cl/assets/media
     networks:
       - cl_net_overlay
-    environment:
-       AWS_ACCESS_KEY_ID: ${AWS_ACCESS_KEY_ID}
-       AWS_SECRET_ACCESS_KEY: ${AWS_SECRET_ACCESS_KEY}
+    env_file:
+      - ../../.env.dev
 
   cl-selenium:
     container_name: cl-selenium
